Added python script to convert blastn (tab).txt to a BED file
Converts the option 6 of blastn output, the tabular .txt file containing the list of hits. Converts the file into 2 new files, 1 BED file which contains the hit region name, hit region start nucleotide and hit region end nucleotide, as well as an RGB colouring system. These files can be used in genome viewers such as Savant. the 2nd file which is optional, contains the average coverage % of all of the contigs which have hits. This txt file can be used to identify the strongest candidates amongst the contigs.